+++ ios/engine/KMEI/KeymanEngine/Classes/Constants.swift
@@ -68,6 +68,11 @@ public enum Defaults {
                                                    font: font,
                                                    oskFont: nil,
                                                    isCustom: false)
+  public static let lexicalModel = InstallableLexicalModel(id: "nrc.en.mtnt",
+                                                   name: "English dictionary (MTNT)",
+                                                   languageID: "en",
+                                                   version: "0.1.2",
+                                                   isCustom: false)
 }
 
 public enum Resources {